Move mail-service into server/src/domain/services
